chore: move wallet slice to slices folder
