+++ components/brave_wallet_ui/common/actions/wallet_actions.ts
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this file,
  * You can obtain one at http://mozilla.org/MPL/2.0/. */
 
- import { WalletActions } from '../reducers/wallet.slice'
+ import { WalletActions } from '../slices/wallet.slice'
 
  // We must re-export actions here until we remove all imports of this file
  export const {
