+++ components/brave_wallet_ui/common/hooks/swap.test.tsx
@@ -14,7 +14,7 @@ import useSwap from './swap'
 // Redux
 import { Provider } from 'react-redux'
 import { combineReducers, createStore } from 'redux'
-import { createWalletReducer } from '../reducers/wallet.slice'
+import { createWalletReducer } from '../slices/wallet.slice'
 import { createPageReducer } from '../../page/reducers/page_reducer'
 
 // Mocks
