+++ components/brave_wallet_ui/page/reducers/index.ts
@@ -6,7 +6,7 @@
 import { combineReducers } from 'redux'
 
 import pageReducer from './page_reducer'
-import walletReducer from '../../common/reducers/wallet.slice'
+import walletReducer from '../../common/slices/wallet.slice'
 import sendCryptoReducer from '../../common/reducers/send_crypto_reducer'
 import accountsTabReducer from './accounts-tab-reducer'
