+++ components/brave_wallet_ui/panel/reducers/index.ts
@@ -6,7 +6,7 @@
 import { combineReducers } from 'redux'
 
 import panelReducer from './panel_reducer'
-import walletReducer from '../../common/reducers/wallet.slice'
+import walletReducer from '../../common/slices/wallet.slice'
 import sendCryptoReducer from '../../common/reducers/send_crypto_reducer'
 
 export default combineReducers({
