+++ components/brave_wallet_ui/stories/wallet-extension-panels.tsx
@@ -58,7 +58,7 @@ import { PanelTitles } from '../options/panel-titles'
 import './locale'
 import { LibContext } from '../common/context/lib.context'
 import { createSendCryptoReducer } from '../common/reducers/send_crypto_reducer'
-import { createWalletReducer } from '../common/reducers/wallet.slice'
+import { createWalletReducer } from '../common/slices/wallet.slice'
 import { createPageReducer } from '../page/reducers/page_reducer'
 import { createPanelReducer } from '../panel/reducers/panel_reducer'
